third_party/WebKit/LayoutTests/external/wpt/html/semantics/tabular-data/processing-model-1/col-span-limits-expected.txt

This is a testharness.js-based test.
PASS col span of 1000 must work 
FAIL col span of 1001 must be treated as 1000 assert_equals: table3 width expected 51 but got 53
Harness: the test ran to completion.
